rename file to command-utils
